requirements.txt
Flask
flask-restful
groq
requests
python-dotenv